Move func VerifySig() to the crypto package
see #1325
